+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Spludo is copyright 2009-2011 by DracoBlue <http://dracoblue.net>
 
 ## 1.1.0-dev
 
+* added bootstrap_manager.whenLoaded(callback) to get notified as
+  soon as the app is ready.
 * replaced self with that
 * fixed core_dev_change_build_version on MACOSX
 * added jshint instead of jslint